requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
controlnet-aux>=0.0.6
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
gradio>=4.0.0
xformers>=0.0.22
safetensors>=0.4.0
omegaconf>=2.3.0
wandb>=0.16.0
scikit-learn>=1.3.0
